test: split test_tools_load_recipe.py into 5 per-concern files
Split the 799-line test_tools_load_recipe.py (11 classes + 1 module-level fn
plus 1 module-level helper and 1 module-level constant) into:

- test_tools_load_recipe.py (kept, trimmed ~270 LoC): core load_recipe tests
  (skill-name rejection, kitchen-gated loading, diagram response, ingredients_only,
  no-context handling) + module-level _MINIMAL_RECIPE_FOR_DIAGRAM constant
- test_load_recipe_contract_validation.py (~95 LoC):
  TestContractMigrationAdapterValidate, TestLoadRecipeSurfacesValidationFailure
- test_load_recipe_exception_handling.py (~190 LoC):
  TestLoadRecipeExceptionHandling, TestLoadRecipeFailClosed
- test_load_recipe_migration.py (~150 LoC):
  TestMigrationSuppression, TestApplyTriageGate
- test_load_recipe_authority.py (~125 LoC):
  TestLoadRecipeReadOnly, TestLoadRecipeAuthorityClobber

The 4-line autouse fixture is copied verbatim into each per-concern file
(per CLAUDE.md Do Not Over-Engineer — extracting one 2-line assignment body
to a helper module plus an import-in-each-file wrapper would add indirection
without simplification).

Mechanical lift: no production code changes, no behavior changes, no test
assertion changes. Each file carries pytestmark = [layer("server"), small]
and the autouse fixture _default_recipe_names_do_not_resolve_as_skills.

Refs: #4610
