+++ tests/server/test_tools_load_recipe.py
@@ -1,23 +1,29 @@
-"""Tests for autoskillit server load_recipe tool."""
+"""Tests for autoskillit server load_recipe tool (kept file).
+
+Carries the core load_recipe tool tests (skill-name rejection, kitchen-gated
+loading, diagram response, ingredients_only, no-context handling) plus the
+shared 4-line autouse fixture. Class-specific behaviors live in the
+per-concern files:
+- test_load_recipe_contract_validation.py — TestContractMigrationAdapterValidate,
+  TestLoadRecipeSurfacesValidationFailure
+- test_load_recipe_exception_handling.py — TestLoadRecipeExceptionHandling,
+  TestLoadRecipeFailClosed
+- test_load_recipe_migration.py — TestMigrationSuppression, TestApplyTriageGate
+- test_load_recipe_authority.py — TestLoadRecipeReadOnly,
+  TestLoadRecipeAuthorityClobber
+"""
 
 from __future__ import annotations
 
 import json
 from pathlib import Path
-from unittest.mock import AsyncMock, MagicMock, patch
+from unittest.mock import MagicMock, patch
 
 import pytest
 
-from autoskillit.config import AutomationConfig
-from autoskillit.core import SkillResolver, SkillResult
-from autoskillit.core.types import RetryReason
-from autoskillit.pipeline.gate import DefaultGateState
+from autoskillit.core import SkillResolver
 from autoskillit.server.tools.tools_recipe import load_recipe
-from tests.server._helpers import (
-    _MINIMAL_SCRIPT_YAML,
-    _resolve_recipe_section,
-    _with_finalized_projection,
-)
+from tests.server._helpers import _resolve_recipe_section
 
 pytestmark = [pytest.mark.layer("server"), pytest.mark.small]
 
@@ -58,14 +64,6 @@ async def test_load_recipe_rejects_skill_names_before_recipe_work(
         assert "ambiguous" in result["error"].lower()
     mock_ctx.recipes.load.assert_not_called()
     mock_ctx.recipes.load_and_validate.assert_not_called()
-
-
-def _write_minimal_script(scripts_dir: Path, name: str = "test-script") -> Path:
-    """Write a minimal valid workflow script with no autoskillit_version field."""
-    scripts_dir.mkdir(parents=True, exist_ok=True)
-    path = scripts_dir / f"{name}.yaml"
-    path.write_text(_MINIMAL_SCRIPT_YAML)
-    return path
 
 
 # ---------------------------------------------------------------------------
@@ -201,328 +199,6 @@ class TestLoadRecipeTools:
         assert findings[0]["message"] == "Invalid recipe structure: injected crash"
 
 
-class TestContractMigrationAdapterValidate:
-    """P7-2: ContractMigrationAdapter.validate uses _load_yaml, not yaml.safe_load."""
-
-    def test_valid_contract_returns_true(self, tmp_path: Path) -> None:
-        from autoskillit.migration.engine import ContractMigrationAdapter
-
-        f = tmp_path / "contract.yaml"
-        f.write_text("skill_hashes:\n  my-skill: abc123\n")
-        adapter = ContractMigrationAdapter()
-        ok, msg = adapter.validate(f)
-        assert ok is True
-        assert msg == ""
-
-    def test_missing_skill_hashes_returns_false(self, tmp_path: Path) -> None:
-        from autoskillit.migration.engine import ContractMigrationAdapter
-
-        f = tmp_path / "contract.yaml"
-        f.write_text("other_field: value\n")
-        adapter = ContractMigrationAdapter()
-        ok, msg = adapter.validate(f)
-        assert ok is False
-        assert "skill_hashes" in msg
-
-    def test_invalid_yaml_returns_false(self, tmp_path: Path) -> None:
-        from autoskillit.migration.engine import ContractMigrationAdapter
-
-        f = tmp_path / "contract.yaml"
-        f.write_bytes(b":\tbad: yaml: [unclosed\n")
-        adapter = ContractMigrationAdapter()
-        ok, msg = adapter.validate(f)
-        assert ok is False
-        assert msg != ""
-
-    def test_missing_file_returns_false(self, tmp_path: Path) -> None:
-        from autoskillit.migration.engine import ContractMigrationAdapter
-
-        adapter = ContractMigrationAdapter()
-        ok, msg = adapter.validate(tmp_path / "nonexistent.yaml")
-        assert ok is False
-        assert msg != ""
-
-
-class TestLoadRecipeExceptionHandling:
-    """CC-1: Outer except in load_recipe must catch anticipated exceptions only."""
-
-    @pytest.fixture(autouse=True)
-    def _setup_ctx(self, tool_ctx_kitchen_open):
-        """Initialize ToolContext with gate open so load_recipe can call _get_config()."""
-
-    @pytest.mark.anyio
-    async def test_yaml_error_surfaces_as_suggestion(
-        self, tmp_path: Path, monkeypatch: pytest.MonkeyPatch
-    ) -> None:
-        """yaml.YAMLError is caught and returned as an error suggestion."""
-        from autoskillit.core.io import YAMLError
-
-        monkeypatch.chdir(tmp_path)
-        recipes_dir = tmp_path / ".autoskillit" / "recipes"
-        recipes_dir.mkdir(parents=True)
-        (recipes_dir / "test.yaml").write_text("name: test\n")
-        with patch(
-            "autoskillit.recipe._api.load_recipe_dict_with_declarations",
-            side_effect=YAMLError("bad yaml"),
-        ):
-            result = json.loads(await load_recipe(name="test"))
-        assert "error" not in result
-        assert any(
-            s.get("rule") == "validation-error" and s.get("severity") == "error"
-            for s in result["suggestions"]
-        )
-
-    @pytest.mark.anyio
-    async def test_value_error_surfaces_as_suggestion(
-        self, tmp_path: Path, monkeypatch: pytest.MonkeyPatch
-    ) -> None:
-        """ValueError (malformed recipe structure) is caught and returned as error suggestion."""
-        from autoskillit.core.types import RecipeSource
-        from autoskillit.recipe.schema import RecipeInfo
-
-        monkeypatch.chdir(tmp_path)
-        recipes_dir = tmp_path / ".autoskillit" / "recipes"
-        recipes_dir.mkdir(parents=True)
-        recipe_path = recipes_dir / "test.yaml"
-        recipe_path.write_text(
-            "name: test\ndescription: Test\nsteps:\n  done:\n    action: stop\n    message: Done\n"
-        )
-        fake_match = RecipeInfo(
-            name="test",
-            description="Test",
-            source=RecipeSource.PROJECT,
-            path=recipe_path,
-        )
-        with (
-            patch("autoskillit.recipe.find_recipe_by_name", return_value=fake_match),
-            patch(
-                "autoskillit.recipe._api._parse_recipe", side_effect=ValueError("bad structure")
-            ),
-        ):
-            result = json.loads(await load_recipe(name="test"))
-        assert "error" not in result
-        assert any(
-            s.get("rule") == "validation-error" and s.get("severity") == "error"
-            for s in result["suggestions"]
-        )
-
-    @pytest.mark.anyio
-    async def test_file_not_found_surfaces_as_suggestion(
-        self, tmp_path: Path, monkeypatch: pytest.MonkeyPatch
-    ) -> None:
-        """FileNotFoundError is caught and returned as an error suggestion."""
-        monkeypatch.chdir(tmp_path)
-        recipes_dir = tmp_path / ".autoskillit" / "recipes"
-        recipes_dir.mkdir(parents=True)
-        (recipes_dir / "test.yaml").write_text(
-            "name: test\ndescription: Test\nsteps:\n  done:\n    action: stop\n    message: Done\n"
-        )
-        with patch(
-            "autoskillit.recipe._api.load_recipe_card",
-            side_effect=FileNotFoundError("missing"),
-        ):
-            result = json.loads(await load_recipe(name="test"))
-        assert "error" not in result
-        assert any(
-            s.get("rule") == "validation-error" and s.get("severity") == "error"
-            for s in result["suggestions"]
-        )
-
-    @pytest.mark.anyio
-    async def test_unexpected_exception_returns_structured_error(
-        self, tmp_path: Path, monkeypatch: pytest.MonkeyPatch
-    ) -> None:
-        """Unexpected exceptions are caught by the handler-level exception boundary."""
-        monkeypatch.chdir(tmp_path)
-        recipes_dir = tmp_path / ".autoskillit" / "recipes"
-        recipes_dir.mkdir(parents=True)
-        (recipes_dir / "test.yaml").write_text(
-            "name: test\ndescription: Test\nsteps:\n  done:\n    action: stop\n    message: Done\n"
-        )
-        with patch(
-            "autoskillit.recipe._api.run_semantic_rules",
-            side_effect=AttributeError("programming error"),
-        ):
-            result = json.loads(await load_recipe(name="test"))
-        assert result["success"] is False
-        assert "error" in result
-        assert "programming error" in result["error"]
-
-
-class TestMigrationSuppression:
-    """SUP1, SUP4: load_recipe respects migration.suppressed config."""
-
-    # SUP1
-    @pytest.mark.anyio
-    async def test_outdated_version_not_in_suggestions_when_suppressed(
-        self, tmp_path, monkeypatch, tool_ctx_kitchen_open
-    ):
-        """SUP1: outdated-recipe-version absent when recipe is suppressed; headless not called."""
-        from autoskillit.config import MigrationConfig
-
-        monkeypatch.chdir(tmp_path)
-        scripts_dir = tmp_path / ".autoskillit" / "recipes"
-        _write_minimal_script(scripts_dir, "test-script")
-
-        tool_ctx_kitchen_open.config = AutomationConfig(
-            migration=MigrationConfig(suppressed=["test-script"])
-        )
-
-        mock_headless = AsyncMock(
-            return_value=SkillResult(
-                success=True,
-                result="ok",
-                session_id="",
-                subtype="success",
-                is_error=False,
-                exit_code=0,
-                needs_retry=False,
-                retry_reason=RetryReason.NONE,
-                stderr="",
-            )
-        )
-        with patch("autoskillit.execution.headless.run_headless_core", mock_headless):
-            result = json.loads(await load_recipe(name="test-script"))
-
-        assert "suggestions" in result
-        rules = [s["rule"] for s in result["suggestions"]]
-        assert "outdated-recipe-version" not in rules
-        mock_headless.assert_not_called()
-
-    # SUP4
-    @pytest.mark.anyio
-    async def test_validate_always_includes_outdated_version_regardless_of_suppression(
-        self, tmp_path, tool_ctx_kitchen_open
-    ):
-        """SUP4: validate_recipe includes outdated-script-version even when suppressed."""
-        from autoskillit.config import MigrationConfig
-        from autoskillit.server.tools.tools_recipe import validate_recipe
-
-        script = tmp_path / "test-script.yaml"
-        script.write_text(_MINIMAL_SCRIPT_YAML + 'autoskillit_version: "0.0.1"\n')
-
-        # Even with script suppressed in config, validate_recipe does not filter
-        tool_ctx_kitchen_open.config = AutomationConfig(
-            migration=MigrationConfig(suppressed=["test-script"])
-        )
-
-        result = json.loads(await validate_recipe(script_path=str(script)))
-        assert "findings" in result
-        rules = [s["rule"] for s in result["findings"]]
-        assert "outdated-recipe-version" in rules
-
-
-class TestApplyTriageGate:
-    """T3: _apply_triage_gate caches triage result and skips on second call."""
-
-    @pytest.fixture(autouse=True)
-    def _close_kitchen(self, tool_ctx):
-        tool_ctx.gate = DefaultGateState(enabled=False)
-
-    @pytest.mark.anyio
-    async def test_apply_triage_gate_second_call_skips_triage(
-        self, tmp_path, monkeypatch, tool_ctx
-    ):
-        """Second _apply_triage_gate call reads from cache; triage_staleness not re-invoked."""
-        import copy
-
-        from autoskillit.recipe.staleness_cache import read_staleness_cache
-        from autoskillit.server._misc import _apply_triage_gate
-
-        monkeypatch.chdir(tmp_path)
-        recipes_dir = tmp_path / ".autoskillit" / "recipes"
-        recipes_dir.mkdir(parents=True)
-        recipe_yaml = (
-            "name: triage-test\ndescription: T\n"
-            "steps:\n  done:\n    action: stop\n    message: Done\n"
-        )
-        recipe_path = recipes_dir / "triage-test.yaml"
-        recipe_path.write_text(recipe_yaml)
-
-        name = "triage-test"
-        result_template = {
-            "content": recipe_yaml,
-            "suggestions": [
-                {
-                    "rule": "stale-contract",
-                    "reason": "hash_mismatch",
-                    "skill": "investigate",
-                    "stored_value": "sha256:old",
-                    "current_value": "sha256:new",
-                    "message": "investigate SKILL.md changed",
-                    "severity": "info",
-                }
-            ],
-            "valid": True,
-        }
-
-        # Get real recipe_info before mocking find
-        recipe_info = tool_ctx.recipes.find(name, Path.cwd())
-        assert recipe_info is not None
-
-        # Mock _ctx.recipes.find to verify it is NOT called when recipe_info is injected
-        mock_find = AsyncMock(return_value=recipe_info)
-        monkeypatch.setattr(tool_ctx.recipes, "find", mock_find)
-
-        mock_triage = AsyncMock(
-            return_value=[{"meaningful": False, "summary": "ok", "skill": "investigate"}]
-        )
-        with patch("autoskillit.server._misc.triage_staleness", mock_triage):
-            # First call: triage_staleness invoked once
-            await _apply_triage_gate(copy.deepcopy(result_template), name, recipe_info=recipe_info)
-
-        assert mock_triage.call_count == 1
-        assert mock_find.call_count == 0, "find() must not be called when recipe_info is injected"
-
-        cache_path = tmp_path / ".autoskillit" / "temp" / "recipe_staleness_cache.json"
-        cached = read_staleness_cache(cache_path, name)
-        assert cached is not None
-        assert cached.triage_result == "cosmetic"
-
-        with patch("autoskillit.server._misc.triage_staleness", mock_triage):
-            # Second call: must read from cache and skip triage_staleness entirely
-            await _apply_triage_gate(copy.deepcopy(result_template), name, recipe_info=recipe_info)
-
-        assert mock_triage.call_count == 1, (
-            "triage_staleness must not be called on second invocation"
-        )
-
-
-class TestLoadRecipeReadOnly:
-    """P4: load_recipe is strictly read-only — no migration, no contract card generation."""
-
-    @pytest.fixture(autouse=True)
-    def _ensure_ctx(self, tool_ctx_kitchen_open):
-        """Ensure server context is initialized with gate open."""
-
-    @pytest.mark.anyio
-    async def test_load_recipe_does_not_call_migration_engine(self, tmp_path, monkeypatch):
-        """load_recipe must not trigger headless migration even when migrations are applicable."""
-        monkeypatch.chdir(tmp_path)
-        with (
-            patch("autoskillit.migration.loader.applicable_migrations", return_value=["v0.1.0"]),
-            patch("autoskillit.execution.headless.run_headless_core") as mock_headless,
-            patch("autoskillit.recipe.contracts.generate_recipe_card") as mock_gen,
-        ):
-            result = json.loads(await load_recipe(name="implementation"))
-        assert "error" not in result
-        mock_headless.assert_not_called()
-        mock_gen.assert_not_called()
-
-    @pytest.mark.anyio
-    async def test_load_recipe_does_not_auto_generate_contract_card(self, tmp_path, monkeypatch):
-        """load_recipe must not call generate_recipe_card even when no card exists."""
-        monkeypatch.chdir(tmp_path)
-        recipes_dir = tmp_path / ".autoskillit" / "recipes"
-        recipes_dir.mkdir(parents=True)
-        (recipes_dir / "test.yaml").write_text(
-            "name: test\ndescription: Test\nsteps:\n  done:\n    action: stop\n    message: Done\n"
-        )
-        with patch("autoskillit.recipe.contracts.generate_recipe_card") as mock_gen:
-            await load_recipe(name="test")
-        mock_gen.assert_not_called()
-
-
 class TestLoadRecipeDiagram:
     """Tests for diagram field in load_recipe responses (DG-12 through DG-15)."""
 
@@ -598,202 +274,3 @@ class TestLoadRecipeIngredientsOnly:
         assert "orchestration_rules" not in result
         assert "stop_step_semantics" not in result
         assert "suggestions" in result
-
-
-class TestLoadRecipeAuthorityClobber:
-    """load_recipe must emit an authority-clobber warning when a server-authoritative
-    key is overridden by the caller. Enforcement (config-layer wins) already works;
-    this test verifies the *feedback* contract."""
-
-    @pytest.mark.anyio
-    async def test_load_recipe_emits_authority_warning(self, tmp_path, monkeypatch):
-        from unittest.mock import AsyncMock, MagicMock
-
-        from tests.server.conftest import _make_mock_ctx
-
-        monkeypatch.chdir(tmp_path)
-        mock_ctx = _make_mock_ctx()
-        mock_ctx.enable_components = AsyncMock()
-        mock_ctx.recipes = MagicMock()
-        mock_recipe_obj = MagicMock()
-        mock_recipe_obj.steps = {"do": MagicMock()}
-        mock_recipe_obj.ingredients = {"base_branch": MagicMock()}
-        mock_ctx.recipes.load.return_value = mock_recipe_obj
-        mock_ctx.recipes.load_and_validate.return_value = _with_finalized_projection(
-            {
-                "content": "name: demo\nsteps:\n  do:\n    tool: run_cmd\n",
-                "valid": True,
-                "suggestions": [],
-                "diagram": None,
-                "ingredients_table": "--- TABLE ---",
-            }
-        )
-        mock_ctx.recipes.find.return_value = MagicMock(path=tmp_path / "demo.yaml")
-        mock_ctx.recipes.load.return_value = mock_recipe_obj
-        mock_ctx.config.migration.suppressed = []
-        mock_ctx.kitchen_id = "test-kitchen"
-        mock_ctx.config.linux_tracing.log_dir = ""
-
-        with patch(
-            "autoskillit.server.tools.tools_recipe._get_ctx_or_none",
-            return_value=mock_ctx,
-        ):
-            with patch(
-                "autoskillit.server.tools.tools_recipe._require_enabled",
-                return_value=None,
-            ):
-                with patch("autoskillit.server.logger"):
-                    with patch(
-                        "autoskillit.server.tools.tools_recipe.resolve_ingredient_defaults",
-                        return_value={
-                            "base_branch": "develop",
-                            "is_fleet_dispatch": "false",
-                            "dispatch_id": "",
-                        },
-                    ):
-                        from autoskillit.server.tools.tools_recipe import load_recipe
-
-                        result_str = await load_recipe(
-                            name="demo",
-                            overrides={"base_branch": "custom"},
-                        )
-
-        parsed = json.loads(result_str)
-        warnings = parsed.get("warnings") or []
-        matching = [w for w in warnings if "base_branch" in w]
-        assert matching, (
-            f"load_recipe must emit a warning naming base_branch; got warnings={warnings}"
-        )
-        server_value_match = [w for w in warnings if "server value 'develop'" in w]
-        assert server_value_match, (
-            "Authority-clobber warning must confirm config value won — "
-            f"expected \"server value 'develop'\" in warning text; got warnings={warnings}"
-        )
-        caller_value_absent = [w for w in warnings if "server value 'custom'" in w]
-        assert not caller_value_absent, (
-            "Authority-clobber warning must NOT report the caller override as the server value — "
-            f"got warnings={warnings}"
-        )
-
-
-# 1i: load_recipe tool surfaces validation failure
-class TestLoadRecipeSurfacesValidationFailure:
-    """When load_and_validate returns valid=False, the load_recipe tool must include
-    a field indicating the recipe failed validation.
-    """
-
-    @pytest.fixture(autouse=True)
-    def _ensure_ctx(self, tool_ctx_kitchen_open):
-        """Ensure server context is initialized with gate open."""
-
-    @pytest.mark.anyio
-    async def test_load_recipe_surfaces_validation_failure(
-        self, tmp_path: Path, monkeypatch: pytest.MonkeyPatch
-    ) -> None:
-        """When load_and_validate returns valid=False, the response must include
-        a validation_failed indicator so callers know the recipe is invalid.
-        """
-        from autoskillit.recipe._api_cache import _LOAD_CACHE
-
-        monkeypatch.chdir(tmp_path)
-        recipes_dir = tmp_path / ".autoskillit" / "recipes"
-        recipes_dir.mkdir(parents=True)
-        (recipes_dir / "no-steps.yaml").write_text(
-            "name: no-steps\ndescription: Missing steps\nkitchen_rules:\n  - 'rule'\n"
-        )
-
-        _LOAD_CACHE.clear()
-
-        result = json.loads(await load_recipe(name="no-steps"))
-        assert result.get("valid") is False
-        assert result.get("validation_failed") is True, (
-            f"Expected validation_failed=True in response; got keys: {list(result.keys())}"
-        )
-        assert "errors" in result
-        assert len(result["errors"]) > 0
-
-
-class TestLoadRecipeFailClosed:
-    """Fail-closed validation for empty and missing content."""
-
-    @pytest.fixture(autouse=True)
-    def _ensure_ctx(self, tool_ctx_kitchen_open):
-        self.ctx = tool_ctx_kitchen_open
-
-    @pytest.mark.anyio
-    async def test_load_recipe_fail_closed_empty_content(self, monkeypatch, tmp_path):
-        monkeypatch.chdir(tmp_path)
-        from autoskillit.recipe._api_cache import _LOAD_CACHE
-
-        _LOAD_CACHE.clear()
-        test_result = {"valid": True, "content": "", "dispatch_feasible": True}
-        monkeypatch.setattr(self.ctx.recipes, "load_and_validate", lambda *a, **kw: test_result)
-        recipe_info = MagicMock(path=tmp_path / "test-recipe.yaml")
-        monkeypatch.setattr(self.ctx.recipes, "find", lambda *a, **kw: recipe_info)
-        monkeypatch.setattr(
-            self.ctx.recipes,
-            "load",
-            lambda *_a, **_kw: MagicMock(steps={}, ingredients={}),
-        )
-        with patch(
-            "autoskillit.server.tools.tools_recipe._apply_triage_gate",
-            new=AsyncMock(return_value=test_result),
-        ):
-            raw = await load_recipe(name="test-recipe")
-        parsed = json.loads(raw)
-        assert parsed["success"] is False
-        assert "content" in parsed["error"].lower()
-
-    @pytest.mark.anyio
-    async def test_load_recipe_fail_closed_missing_content(self, monkeypatch, tmp_path):
-        monkeypatch.chdir(tmp_path)
-        from autoskillit.recipe._api_cache import _LOAD_CACHE
-
-        _LOAD_CACHE.clear()
-        test_result = {"valid": True, "dispatch_feasible": True}
-        monkeypatch.setattr(self.ctx.recipes, "load_and_validate", lambda *a, **kw: test_result)
-        recipe_info = MagicMock(path=tmp_path / "test-recipe.yaml")
-        monkeypatch.setattr(self.ctx.recipes, "find", lambda *a, **kw: recipe_info)
-        monkeypatch.setattr(
-            self.ctx.recipes,
-            "load",
-            lambda *_a, **_kw: MagicMock(steps={}, ingredients={}),
-        )
-        with patch(
-            "autoskillit.server.tools.tools_recipe._apply_triage_gate",
-            new=AsyncMock(return_value=test_result),
-        ):
-            raw = await load_recipe(name="test-recipe")
-        parsed = json.loads(raw)
-        assert parsed["success"] is False
-        assert "content" in parsed["error"]
-
-    @pytest.mark.anyio
-    async def test_load_recipe_fail_closed_missing_finalized_projection(
-        self,
-        monkeypatch,
-        tmp_path,
-    ):
-        monkeypatch.chdir(tmp_path)
-        test_result = {
-            "valid": True,
-            "content": "name: test-recipe\n",
-            "dispatch_feasible": True,
-        }
-        monkeypatch.setattr(self.ctx.recipes, "load_and_validate", lambda *a, **kw: test_result)
-        recipe_info = MagicMock(path=tmp_path / "test-recipe.yaml")
-        monkeypatch.setattr(self.ctx.recipes, "find", lambda *a, **kw: recipe_info)
-        monkeypatch.setattr(
-            self.ctx.recipes,
-            "load",
-            lambda *_a, **_kw: MagicMock(steps={}, ingredients={}),
-        )
-        with patch(
-            "autoskillit.server.tools.tools_recipe._apply_triage_gate",
-            new=AsyncMock(return_value=test_result),
-        ):
-            raw = await load_recipe(name="test-recipe")
-
-        parsed = json.loads(raw)
-        assert parsed["success"] is False
-        assert "finalized projection" in parsed["error"]
